requirements.txt
Flask
PyYAML
gpiozero
rpi-lgpio
picamera2
Pillow
imageio
inky[rpi,fonts]
opencv-python
